requirements.txt
cozepy
openai
pyaudio
pygame
numpy
baidu-aip
dotenv
flask